+++ pkg/hyperpic/provider/memory/configuration.go
@@ -10,7 +10,7 @@ import "time"
 type CacheConfiguration struct {
 	LifeTime      time.Duration `mapstructure:"life_time"`
 	CleanInterval time.Duration `mapstructure:"clean_interval"`
-	MemoryLimit   int           `mapstructure:"memory_limit"`
+	MemoryLimit   uint64        `mapstructure:"memory_limit"`
 }
 
 // SourceConfiguration struct.
